% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_memes.R
\name{get_memes}
\alias{get_memes}
\title{Get memes}
\usage{
get_memes()
}
\value{
A data frame with one row per meme.
}
\description{
Gets a data frame of popular memes that may be captioned with the Imgflip
API. The size of this array and the order of memes may change at any time.
}
\examples{
\dontrun{
memes <- get_memes()
id <- memes[memes$name == "Batman Slapping Robin", "id"]
id
}
}
